Incorporate python api blog post feedback from Chuck and Vid
The commit incorporates the feedback from Vid and Chuck on the rough
brainstorm of ideas I put together initially. The first few paragraphs
are based on a suggested opening from Vid.

The goal here is to motivate a Python programmer who has heard of
stream processing to download Wallaroo and check it out. The post
explains the problems of distributed stream processing, and shows that
we have created an API that makes it is to use Python to build a
scale-agnostic stream processing application.