Replace double-starts with backticks in the Twitter Hashtag post
Parts of the blog post use double-stars to surround code, which is
rendered as bold in some markdown engines. Hugo doesn't support the
use of "**" for bold, and our standard style is to use backticks for
code, so this change makes the text display correctly and consistently
with our other posts.