Move pip instructions, change compute method description
The pip instructions were at the end of the blog post. They take a
long time to complete, so they were moved to the beginning so that the
user can continue with other instructions while the packages are
downloading.

There was a section that described the compute method, but did not
describe compute_multi, which was used by the class below it. The
description has been updated to describe both compute and
compute_multi.

Changed formatting of some of the headers because they used numbers,
which is inconsistent with our other posts and didn't really add
anything to the readability of the post.