Add "you might also like" section to blog posts
Adds a "you might also like" list of 3 blog posts that are related to the post the reader is viewing based on the intersection of tags. Posts with 2 or more tags in common are favored, those with 1 or more in common are also in play for being displayed. The top 3 posts are selected and then displayed in random order.